Add user profile front end + add Model for future display usage between
form or detail pages

+++ server/routes/utils/baseresource.py
@@ -1,31 +1,52 @@
 from .. import Resource, request, IntegrityError, ValidationError
+from ..helpers.query_helpers import get_all, get_all_by_condition, get_instance_by_id, get_one_by_condition
+
+from models.user import User
 
 from config import db
+import ipdb
 
 class BaseResource(Resource):
     model = None
     schema = None
 
     def get(self, id=None, condition=None):
+        
         try:
             if id is None and condition is None:
-                instances = self.model.query.all()
-                return self.schema.dump(instances, many=True), 200
+                instances = get_all(self.model)
+                
+                return (
+                    self.schema.dump(instances, many=True),
+                    200,
+                )  # Use the schema to serialize the instances
             elif condition is not None:
-                instances = self.model.query.filter_by(**condition).all()
-                return self.schema.dump(instances, many=True), 200
+                instances = get_all_by_condition(self.model, condition)
+                
+
+                return (
+                    self.schema.dump(instances, many=True),
+                    200,
+                )  # Use the schema to serialize the instances
             else:
-                instance = self.model.query.get(id)
+                instance = get_instance_by_id(self.model, id)
                 if instance is None:
                     return {"errors": f"{self.model.__name__} not found"}, 404
-                return self.schema.dump(instance), 200
+                
+                return (
+                    self.schema.dump(instance),
+                    200,
+                )  # Use the schema to serialize the instance
         except Exception as e:
             db.session.rollback()
             return {"errors": str(e)}, 400
 
     def delete(self, id=None):
         try:
-            instance = self.model.query.get(id)
+            instance = get_instance_by_id(self.model, id)
+            if instance is None:
+                return {"errors": f"{self.model.__name__} not found"}, 404
+
             db.session.delete(instance)
             db.session.commit()
             return "", 204
@@ -47,13 +68,30 @@ class BaseResource(Resource):
             db.session.rollback()
             return {"message": "Invalid data"}, 422
 
-    def patch(self, id=None):
+    def patch(self, id):
         try:
+            ipdb.set_trace()
+
             data = self.schema.load(request.json)
-            instance = self.model.query.get(id)
+                # Check if data is a User instance (object not dictionary)
+            if isinstance(data, User):
+                ipdb.set_trace()
+
+                # Convert the User instance to a dictionary
+                data = {c.name: getattr(data, c.name) for c in data.__table__.columns}
+            ipdb.set_trace()
+
+            instance = get_instance_by_id(self.model, id)
+            if instance is None:
+                return {"errors": f"{self.model.__name__} not found"}, 404
+            ipdb.set_trace()
+
             for key, value in data.items():
-                setattr(instance, key, value)
+                if value is not None:
+                    setattr(instance, key, value)
             db.session.commit()
+            ipdb.set_trace()
+
             return self.schema.dump(instance), 200
         except ValidationError as e:
             return {"message": str(e)}, 422
